refactor: challengeDetailPage custom hook 분리
